Allow TYPE parameter on X- properties
According to the spec, the TYPE parameter MUST not be applied on
properties defined in the spec, but not explicitly listed in section
5.6 of the RFC.

https://www.rfc-editor.org/rfc/rfc6350#section-5.6

This means that the parameter MAY be applied on properties not
explicitly listed in the spec, such as non-standard 'X-' properties.

The patch allows TYPE parameter on such properties.