Add Uri new type.
This allows us to implement FromStr for the new type and therefore use
serde_with::serde_as to serialize Uri values to strings.

The default serde implementation for uriparse is rather verbose and
unnecessary.

As we are using serde_as in conjunction with cfg_attr it is required to
add the cfg_eval dependency so the macros are expanded correctly.